test: serialization: Prevent parallelisation
The ipa_data_serializer_test is configured to be able to run in parallel
with other tests.

If the test tries to acquire the VIMC sensor B while it is in use, it
would skip require skipping the test.

Instead, inform meson that it can not be run in parallel to ensure that
the test is able to obtain the resources it requires for execution.

Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Jean-Michel Hautbois <jeanmichel.hautbois@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>