README: Document dependencies for backtrace generation
libcamera now has the ability to use libdw and libunwind to generate
backtraces, in addition to the glibc backtrace() function. libdw
provides the most detailed output and is highly recommended, but is
limited to parsing backtraces, it doesn't support capturing them.
libunwind and backtrace() provide both features. If backtrace() is
available, libunwind will not bring any improvement.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Umang Jain <umang.jain@ideasonboard.com>